Pin deps and document CLI
