+++ requirements.txt
@@ -1,9 +1,10 @@
-yfinance
-pandas
-numpy
-ib_insync
-pandas_datareader
-requests
-rich
-pypdf
-fpdf
+yfinance==0.2.64
+pandas==2.3.0
+numpy==2.3.1
+ib_insync==0.9.86
+pandas_datareader==0.10.0
+requests==2.32.4
+urllib3==2.5.0
+rich==14.0.0
+pypdf==5.7.0
+fpdf==1.7.2
